chore: move common date picker code to common file
